# api/openapi-spec/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple REST API for performing basic arithmetic operations",
        "title": "Calculator REST API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/multiply": {
            "post": {
                "description": "Calculate the multiply of numbers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calculator"
                ],
                "summary": "Multiply of numbers",
                "parameters": [
                    {
                        "description": "Data Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.DataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.MultiplyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.BadRequest"
                        }
                    }
                }
            }
        },
        "/sum": {
            "post": {
                "description": "Calculate the sum of a list of numbers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calculator"
                ],
                "summary": "Sum of numbers",
                "parameters": [
                    {
                        "description": "Data Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.DataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SumResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.BadRequest"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.BadRequest": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Bad request"
                }
            }
        },
        "main.DataRequest": {
            "type": "object",
            "properties": {
                "numbers": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        1.5,
                        2,
                        3
                    ]
                },
                "uuid": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
        "main.MultiplyResponse": {
            "type": "object",
            "properties": {
                "multiply": {
                    "type": "number",
                    "example": 6.1
                },
                "uuid": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
        "main.SumResponse": {
            "type": "object",
            "properties": {
                "sum": {
                    "type": "number",
                    "example": 6.3
                },
                "uuid": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        }
    }
}